Update Feedback, Invoice

+++ Koi-Veterinary-Service-Center-API/src/main/java/com/fpt/Koi_Veterinary_Service_Center_API/dto/response/reportResponse.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class reportResponse {
     private int totalOrders;
     private int totalCustomers;
-    private float totalPayments;
+    private long totalPayments;
     private String bestService;
 }
