+++ Koi-Veterinary-Service-Center-API/src/main/java/com/fpt/Koi_Veterinary_Service_Center_API/service/impl/ReportServiceImpl.java
@@ -43,7 +43,7 @@ public class ReportServiceImpl implements IReportService {
         }
 
         //get all invoices total
-        float total = 0;
+        long total = 0;
         for(Invoice invoice: invoices){
             total += invoice.getTotal();
         }
